connecting jobs and users

+++ routes/jobs.js
@@ -5,6 +5,7 @@ const Jobs = require('../models/job');
 const fsExtra = require('fs-extra');
 const dirPath = 'public/resumes';
 var authenticate = require('../authenticate');
+const Users = require('../models/user');
 
 const cors = require('./cors');
 
@@ -33,7 +34,7 @@ const fileUpload = express.Router();
 fileUpload.route('/')
 .options(cors.corsWithOptions, (req, res) => { res.sendStatus(200); })
 .get(cors.cors, authenticate.verifyUser, (req, res, next) => {
-    Jobs.find({})
+    Jobs.find({userID: req.user._id})
     .then((jobs) => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -42,6 +43,7 @@ fileUpload.route('/')
     .catch((err) => next(err));
 })
 .post(cors.corsWithOptions, authenticate.verifyUser, upload.array('filefield', 10000), (req, res, next) => {
+    req.body.userID = req.user._id;
     Jobs.create(req.body)
     .then((job) => {
         req.files.map((file) => {
@@ -56,6 +58,7 @@ fileUpload.route('/')
         .catch((err) => next(err));
     }, (err) => next(err))
     .catch((err) => next(err));
+
 })
 .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
@@ -76,13 +79,24 @@ fileUpload.route('/')
 });
 
 fileUpload.route('/:jobID')
-.options(cors.corsWithOptions, authenticate.verifyUser, (req, res) => { res.sendStatus(200); })
-.get(cors.cors, (req, res, next) => {
+.options(cors.corsWithOptions, (req, res) => { res.sendStatus(200); })
+.get(cors.cors, authenticate.verifyUser, (req, res, next) => {
     Jobs.findById(req.params.jobID)
     .then((job) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(job);
+        //unautherized
+        if(req.user._id != job.userID)
+        {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({sucess: "false", msg: "unautherized"});
+        }
+        //autherized
+        else
+        {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(job);
+        }
     }, (err) => next(err))
     .catch((err) => next(err));
 })
@@ -93,28 +107,58 @@ fileUpload.route('/:jobID')
 })
 
 .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
-    Jobs.findByIdAndUpdate(req.params.jobID,
-                           {$set: req.body},
-                           {new: true})
-
+    Jobs.findById(req.params.jobID)
     .then((job) => {
-        //nlp model(req.body.description, job.resumes)
-        res.statusCode = 200,
-        res.setHeader('Content-Type', 'application/json'),
-        res.json(job);
+        //unutherized
+        if(req.user._id != job.userID)
+        {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({sucess: "false", msg: "unautherized"});
+        }
+        //autherized
+        else
+        {
+            Jobs.findByIdAndUpdate(req.params.jobID,
+                {$set: req.body},
+                {new: true})
+
+            .then((job) => {
+            //nlp model(req.body.description, job.resumes)
+            res.statusCode = 200,
+            res.setHeader('Content-Type', 'application/json'),
+            res.json(job);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+        }
     }, (err) => next(err))
     .catch((err) => next(err));
 })
 
 .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
-    Jobs.findByIdAndRemove(req.params.jobID)
+    Jobs.findById(req.params.jobID)
     .then((job) => {
-        job.resumes.map((resume) => {
-            fsExtra.removeSync(resume.path);
-        });
-        res.statusCode = 200,
-        res.setHeader('Content-Type', 'application/json');
-        res.json({sucess: "true"});
+        //unautherized
+        if(req.user._id != job.userID)
+        {
+            res.statusCode = 401;
+            res.setHeader('Content-Type', 'application/json');
+            res.json({sucess: "false", msg: "unautherized"});
+        }
+        //autherized
+        else
+        {
+            Jobs.findByIdAndRemove(req.params.jobID)
+            .then((job) => {
+                job.resumes.map((resume) => {
+                    fsExtra.removeSync(resume.path);
+                });
+                res.statusCode = 200,
+                res.setHeader('Content-Type', 'application/json');
+                res.json({sucess: "true", msg: "removed successfully"});
+            }, (err) => next(err))
+            .catch((err) => next(err));
+        }
     }, (err) => next(err))
     .catch((err) => next(err));
 });
@@ -123,7 +167,7 @@ fileUpload.route('/:jobID/resumes')
 .options(cors.corsWithOptions, (req, res) => { res.sendStatus(200); })
 .get(cors.cors, authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
-    res.end('PUT operation not supported on /jobs' + req.params.jobID + '/resumes');
+    res.end('Get operation not supported on /jobs' + req.params.jobID + '/resumes');
 })
 
 .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
@@ -132,20 +176,29 @@ fileUpload.route('/:jobID/resumes')
 })
 
 .post(cors.corsWithOptions, authenticate.verifyUser, upload.array('filefield', 10000), (req, res, next) => {
-    //nlp_model
     Jobs.findById(req.params.jobID)
     .then((job) => {
-        req.files.map((file) => {
-            job.resumes = job.resumes.concat({filename: file.filename, path: file.path, percentage: Math.random(), jobId: job._id});
-        });
-
-        job.save()
-        .then((job) => {
-            res.statusCode = 200;
+        //unautherized
+        if(req.user._id != job.userID)
+        {
+            res.statusCode = 401;
             res.setHeader('Content-Type', 'application/json');
-            res.json(job);
-        }, (err) => next(err))
-        .catch((err) => next(err));
+            res.json({sucess: "false", msg: "unautherized"});
+        }
+        //autherized
+        else
+        {
+            req.files.map((file) => {
+                job.resumes = job.resumes.concat({filename: file.filename, path: file.path, percentage: Math.random(), jobId: job._id});
+            });
+            job.save()
+            .then((job) => {
+                res.statusCode = 200;
+                res.setHeader('Content-Type', 'application/json');
+                res.json(job);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+        }
     }, (err) => next(err))
     .catch((err) => next(err));
 })
@@ -160,18 +213,29 @@ fileUpload.route('/:jobID/resumes/:resumeID')
 .get(cors.cors, authenticate.verifyUser, (req, res, next) => {
     Jobs.findById(req.params.jobID)
     .then((job) => {
-        resume = job.resumes.find(cv => cv._id == req.params.resumeID);
-        if(resume)
+        //unautherized
+        if(req.user._id != job.userID)
         {
-            res.statusCode = 200;
+            res.statusCode = 401;
             res.setHeader('Content-Type', 'application/json');
-            res.json(resume);
+            res.json({sucess: "false", msg: "unautherized"});
         }
+        //autherized
         else
         {
-            res.statusCode = 404;
-            res.setHeader('Content-Type', 'application/json');
-            res.json({sucess: "false", msg: "didn't find this resume"});
+            resume = job.resumes.find(cv => cv._id == req.params.resumeID);
+            if(resume)
+            {
+                res.statusCode = 200;
+                res.setHeader('Content-Type', 'application/json');
+                res.json(resume);
+            }
+            else
+            {
+                res.statusCode = 404;
+                res.setHeader('Content-Type', 'application/json');
+                res.json({sucess: "false", msg: "didn't find this resume"});
+            }
         }
     }, (err) => next(err))
     .catch((err) => next(err));
@@ -190,18 +254,29 @@ fileUpload.route('/:jobID/resumes/:resumeID')
 .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
     Jobs.findById(req.params.jobID)
     .then((job) => {
-        resume = job.resumes.find(cv => cv._id == req.params.resumeID);
-        fsExtra.removeSync(resume.path);
-        job.resumes.remove(resume);
-        job.save()
-        .then((job) => {
-            res.statusCode = 200;
+        //unautherized
+        if(req.user._id != job.userID)
+        {
+            res.statusCode = 401;
             res.setHeader('Content-Type', 'application/json');
-            res.json(job);
-        }, (err) => next(err))
-        .catch((err) => next(err));
-    })
-    .catch((err) => next(err))
+            res.json({sucess: "false", msg: "unautherized"});
+        }
+        //autherized
+        else
+        {
+            resume = job.resumes.find(cv => cv._id == req.params.resumeID);
+            fsExtra.removeSync(resume.path);
+            job.resumes.remove(resume);
+            job.save()
+            .then((job) => {
+                res.statusCode = 200;
+                res.setHeader('Content-Type', 'application/json');
+                res.json(job);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+        }
+    }, (err) => next(err))
+    .catch((err) => next(err));
 });
 
 module.exports = fileUpload;
